[JacobiSet] Rename .inl
